virJSONValueArraySize: return ssize_t
The internal representation of a JSON array counts the items in
size_t. However, for some reason, when asking for the count it's
reported as int. Firstly, we need the function to return a signed
type as it's returning -1 on an error. But, not every system has
integer the same size as size_t. Therefore, lets return ssize_t.

+++ src/network/bridge_driver.c
@@ -3599,7 +3599,7 @@ networkGetDHCPLeases(virNetworkPtr network,
     size_t i, j;
     size_t nleases = 0;
     int rv = -1;
-    int size = 0;
+    ssize_t size = 0;
     int custom_lease_file_len = 0;
     bool need_results = !!leases;
     long long currtime = 0;
